workaround for limited unzip capabilities
As suggested by SuperR.
(http://forum.xda-developers.com/member.php?u=5787964)

Create a tmpunzip directory in the target directory where we unzip the
contents. Then we check if there is a directory in tmpunzip with the
same name as the directory we just extracted. If so, we move the files
from inside that directory to the target. Otherwise we move the contents
of tmpunzip to the target. Once finished, we delete tmpunzip.