merge develop into master (#1)
* include updater-script

We should still include an updater-script, but since it isn't being
parsed we just make it contain a dummy comment.

* update-binary as a shell script

A fairly complete EDIFY script feature replacement using shell functions

* Adding some white space and indentations

Pure cosmetic changes.
It's just easier to read and editors can collapse the functions.

* workaround for limited unzip capabilities

As suggested by SuperR.
(http://forum.xda-developers.com/member.php?u=5787964)

Create a tmpunzip directory in the target directory where we unzip the
contents. Then we check if there is a directory in tmpunzip with the
same name as the directory we just extracted. If so, we move the files
from inside that directory to the target. Otherwise we move the contents
of tmpunzip to the target. Once finished, we delete tmpunzip.